initial refactor of `FormControlLabel` and Storybook story
