+++ packages/manager/src/components/ModeSelect/ModeSelect.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { Radio } from 'src/components/Radio/Radio';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import RadioGroup from 'src/components/core/RadioGroup';
 
 export interface Mode<modes> {
