+++ packages/manager/src/components/Radio/Radio.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
 
-import FormControlLabel from '../core/FormControlLabel';
+import { FormControlLabel } from '../FormControlLabel';
 import RadioGroup from '../core/RadioGroup';
 import { Radio } from './Radio';
