--- packages/manager/src/components/core/FormControlLabel.ts
@@ -1,8 +1,0 @@
-import FormControlLabel, {
-  FormControlLabelProps as _FormControlLabelProps,
-} from '@mui/material/FormControlLabel';
-
-/* tslint:disable-next-line:no-empty-interface */
-export type FormControlLabelProps = _FormControlLabelProps;
-
-export default FormControlLabel;
