+++ packages/manager/src/features/Account/AutoBackups.tsx
@@ -8,7 +8,7 @@ import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 
 const useStyles = makeStyles()((theme: Theme) => ({
   enableBackupsButton: {
