+++ packages/manager/src/features/Backups/AutoEnroll.tsx
@@ -6,7 +6,7 @@ import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import { Paper } from 'src/components/Paper';
 
 type ClassNames = 'header' | 'root' | 'toggleLabel' | 'toggleLabelText';
