+++ packages/manager/src/features/Databases/DatabaseCreate/DatabaseCreate.tsx
@@ -39,7 +39,7 @@ import { RegionHelperText } from 'src/components/SelectRegionPanel/RegionHelperT
 import { TextField } from 'src/components/TextField';
 import { Typography } from 'src/components/Typography';
 import FormControl from 'src/components/core/FormControl';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import { Paper } from 'src/components/Paper';
 import RadioGroup from 'src/components/core/RadioGroup';
 import { databaseEngineMap } from 'src/features/Databases/DatabaseLanding/DatabaseRow';
