+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeSettings/AlertSection.tsx
@@ -7,7 +7,7 @@ import { Divider } from 'src/components/Divider';
 import { TextField } from 'src/components/TextField';
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import { InputAdornment } from 'src/components/InputAdornment';
 import { fadeIn } from 'src/styles/keyframes';
