+++ packages/manager/src/features/NodeBalancers/NodeBalancerPassiveCheck.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import FormHelperText from 'src/components/core/FormHelperText';
 
 import type { NodeBalancerConfigPanelProps } from './types';
