+++ packages/manager/src/features/Profile/AuthenticationSettings/TwoFactor/TwoFactorToggle.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { Toggle } from 'src/components/Toggle';
 import FormControl from 'src/components/core/FormControl';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 
 interface ToggleProps {
   disabled?: boolean;
