+++ packages/manager/src/features/Profile/Settings/Settings.tsx
@@ -6,13 +6,13 @@ import * as React from 'react';
 import { isOSMac } from 'src/App';
 import { Code } from 'src/components/Code/Code';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
+import { FormControlLabel } from 'src/components/FormControlLabel';
+import { Paper } from 'src/components/Paper';
 import { PreferenceToggle } from 'src/components/PreferenceToggle/PreferenceToggle';
 import { Radio } from 'src/components/Radio/Radio';
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
 import FormControl from 'src/components/core/FormControl';
-import FormControlLabel from 'src/components/core/FormControlLabel';
-import { Paper } from 'src/components/Paper';
 import RadioGroup from 'src/components/core/RadioGroup';
 import { useMutatePreferences, usePreferences } from 'src/queries/preferences';
 import { useMutateProfile, useProfile } from 'src/queries/profile';
