+++ packages/manager/src/features/Volumes/VolumeDrawer/ModeSelection.tsx
@@ -2,7 +2,7 @@ import { makeStyles } from '@mui/styles';
 import * as React from 'react';
 
 import { Radio } from 'src/components/Radio/Radio';
-import FormControlLabel from 'src/components/core/FormControlLabel';
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import RadioGroup from 'src/components/core/RadioGroup';
 
 const useStyles = makeStyles({
